LPD-55588 Refactor GDPR tests
