+++ modules/test/playwright/tests/blogs-web/main/pages/BlogsPage.ts
@@ -14,6 +14,7 @@ export class BlogsPage {
 	readonly page: Page;
 	readonly permissionsFrameLocator: FrameLocator;
 	readonly selectAllBlogEntriesCheckBox: Locator;
+	readonly successMessage: Locator;
 
 	constructor(page: Page) {
 		this.blogName = (title: string) => page.getByText(`${title}`);
@@ -27,6 +28,7 @@ export class BlogsPage {
 		this.selectAllBlogEntriesCheckBox = page.getByLabel(
 			'Select All Items on the Page'
 		);
+		this.successMessage = page.getByText('Successful');
 	}
 
 	async goto(siteUrl?: Site['friendlyUrlPath']) {
